requirements.txt
pytest==8.3.4
selenium==4.28.1
allure-pytest==2.13.5